# MemoryKnowledge/openapi.yaml
openapi: 3.0.3
info:
  title: Knowledge Service API
  description: |
    Standalone knowledge service for Code-Graph + LLM-Wiki.
    Provides 28 endpoints for managing and querying code repositories
    and wiki knowledge bases.

    ## Multi-tenancy (001)

    Every endpoint requires the tenant identity `service_id` in the
    **`x-tdai-service-id` request header** (`^[A-Za-z0-9_-]+$`, missing/malformed → 400).
    This is the same value the kernel uses for store/VDB routing, so tenant identity is
    unified end-to-end. All lookups are scoped by `service_id`; an id-only endpoint
    targeting another tenant's resource returns 404. `service_id` is never a request-body
    field and is not echoed in responses.

    ## Asset ID Model

    Assets have globally unique, immutable IDs:

    | Asset Type  | Field           | Prefix   | Example        |
    |-------------|-----------------|----------|----------------|
    | LLM-Wiki    | `wiki_id`       | `wiki-`  | `wiki-9c1f2b`  |
    | Code-Graph  | `code_graph_id` | `cg-`    | `cg-7e3d10`    |

    ## Response Envelope

    All responses use `ApiResponseEnvelope`:
    - `code: 0` means success; non-zero means business error
    - `data` contains the payload on success, `null` on error
    - `message` provides human-readable status

  version: "1.0.0"
  contact:
    name: Knowledge Service
servers:
  - url: http://localhost:8421/v3
    description: Local development server
...
paths:
...
  /wiki/delete:
    post:
      tags: [LLM-Wiki]
      summary: Batch delete wikis
      security:
        - LifecycleBearer: []
      description: |
        Deletes wikis in any status, including `pending`/`processing` (008): a
        delete never returns 409. In-flight ingest is signalled to abort at its
        next checkpoint; the row is hard-deleted and connection/disk/index
        resources are cleaned up. Cross-memory ids report as "not found".
      operationId: deleteWiki
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WikiDeleteRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BatchDeleteResult'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
  /code-graph/delete:
    post:
      tags: [Code-Graph]
      summary: Batch delete code graphs
      security:
        - LifecycleBearer: []
      description: |
        Deletes code graphs in any status, including `pending`/`processing`
        (008): a delete never returns 409. In-flight clone/index is signalled to
        abort at its next checkpoint; the row is hard-deleted and instance
        pool / disk resources are released. Cross-memory ids report as "not found".
      operationId: deleteCodeGraph
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CodeGraphDeleteRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BatchDeleteResult'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
  /internal/lifecycle/ownership/transfer:
    post:
      tags: [Internal Lifecycle]
      summary: Compare-and-swap Wiki or Code Graph backing owner
      operationId: transferKnowledgeOwnershipInternal
      security:
        - LifecycleBearer: []
      parameters:
        - $ref: '#/components/parameters/ServiceIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [resource_type, resource_id, from_owner_user_id, to_owner_user_id]
              properties:
                resource_type:
                  type: string
                  enum: [llm_wiki, code_graph]
                resource_id: { type: string }
                from_owner_user_id: { type: string }
                to_owner_user_id: { type: string }
      responses:
        '200': { description: Backing owner transferred or already at target }
        '401': { description: Invalid lifecycle authorization }
        '409': { description: Owner changed or resource is processing }
        '503': { description: Lifecycle authorization is not configured }
  /internal/lifecycle/integrity/inventory:
    post:
      tags: [Internal Lifecycle]
      summary: List content-free Knowledge inventory for integrity reconciliation
      operationId: listKnowledgeIntegrityInventoryInternal
      security:
        - LifecycleBearer: []
      parameters:
        - $ref: '#/components/parameters/ServiceIdHeader'
      responses:
        '200': { description: Content-free Wiki and Code Graph inventory }
        '401': { description: Invalid lifecycle authorization }
        '503': { description: Lifecycle authorization is not configured }
components:
  securitySchemes:
    LifecycleBearer:
      type: http
      scheme: bearer
      description: KNOWLEDGE_AUTH_TOKEN for Panel control-plane calls only.
  parameters:
    ServiceIdHeader:
      name: x-tdai-service-id
      in: header
      required: true
      description: >-
        Tenant identity `service_id` (001 multi-tenancy). Required on every endpoint;
        same value the kernel uses for store/VDB routing. Missing/malformed → 400.
      schema:
        type: string
        pattern: "^[A-Za-z0-9_-]+$"
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponseEnvelope'
  schemas:
    ApiResponseEnvelope:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          description: 0 = success, non-zero = business error
        message:
          type: string
        request_id:
          type: string
        data:
          nullable: true
...
    WikiDeleteRequest:
      type: object
      required: [wiki_ids]
      properties:
        wiki_ids:
          type: array
          items:
            type: string
          maxItems: 100
...
    CodeGraphDeleteRequest:
      type: object
      required: [code_graph_ids]
      properties:
        code_graph_ids:
          type: array
          items:
            type: string
          maxItems: 100
...
    BatchDeleteResult:
      type: object
      properties:
        deleted_ids:
          type: array
          items:
            type: string
        failed:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              reason:
                type: string